combine modifier/preset test scripts
