--- .github/workflows/test-presets.yml
@@ -1,36 +1,0 @@
-name: 'Test: Preset Data'
-
-on:
-  workflow_dispatch:
-  pull_request:
-    paths:
-      - src/assets/presetdata/**
-
-env:
-  NODE_VERSION: 18
-
-jobs:
-  testPresets:
-    runs-on: ubuntu-latest
-    timeout-minutes: 5
-
-    steps:
-      - name: Checkout 🛎️
-        uses: actions/checkout@v3
-
-      - name: Setup Node
-        uses: actions/setup-node@v3
-        with:
-          node-version: ${{ env.NODE_VERSION }}
-      - name: Cache dependencies
-        id: node-modules-cache
-        uses: actions/cache@v3
-        with:
-          path: node_modules
-          key: ${{ runner.os }}-${{ env.NODE_VERSION }}-node_modules-${{ hashFiles('**/yarn.lock') }}
-      - name: Install dependencies
-        run: yarn install --frozen-lockfile
-        if: steps.node-modules-cache.outputs.cache-hit != 'true'
-
-      - name: Test preset data
-        run: yarn testPresets
