+++ package.json
@@ -25,8 +25,9 @@
     "eslint": "eslint --ignore-path .gitignore --max-warnings 0 . --ext .js,.jsx,.ts,.tsx",
     "updateMappings": "node updateMappings.js",
     "deploy": "vite build && gh-pages -d dist",
-    "testModifiers": "node ./src/assets/testModifiers.js",
-    "testPresets": "node ./src/assets/testPresets.js",
+    "testData": "node ./src/assets/testData.js",
+    "testModifiers": "echo 'this script has been replaced with testData!'",
+    "testPresets": "echo 'this script has been replaced with testData!'",
     "extractLocale": "babel --config-file ./babel-extract.config.js 'src/**/*.{js,jsx,ts,tsx}' > /dev/null",
     "extractLocaleEn": "babel --config-file ./babel-extract-en.config.js 'src/**/*.{js,jsx,ts,tsx}' > /dev/null && node cleanupLocale.js",
     "extractLocaleCMD": "babel --config-file ./babel-extract.config.js \"src/**/*.{js,jsx,ts,tsx}\" > NUL",
